Loading/Generating Precomputed MismatchL2 Values
Major upgrade to <getAllCorrelations.m>, it now accesses a save file of mismatchL2 values, or generates the same for the first use. <displaySCNI.m> has been modified to account for the same, and to avoid conflicts with <analyzeData.m> over usage of this common function. The earlier commit was faulty since it would fail to generate correlation plots due to this incompatibility. Minor modifications elsewhere.